+++ bloop/__init__.py
@@ -26,6 +26,8 @@ from .types import (
     Binary,
     Boolean,
     DateTime,
+    DynamicList,
+    DynamicMap,
     Integer,
     List,
     Map,
@@ -52,8 +54,9 @@ __all__ = [
 
     # Types
     "UUID", "Binary", "Boolean", "DateTime", "Integer", "List", "Map", "Number", "Set", "String", "Timestamp",
+    "DynamicList", "DynamicMap",
 
     # Misc
     "Condition", "QueryIterator", "ScanIterator", "Stream", "missing"
 ]
-__version__ = "2.1.0"
+__version__ = "2.2.0"